+++ resources/js/builder/inventory/components/configuration/button/Messages.tsx
@@ -1,4 +1,5 @@
 import {Form} from "react-bootstrap";
+import AutoCompleteFormControl from "../../utils/AutoCompleteFormControl";
 
 const Messages = ({currentSlot, handleChange}) => {
 
@@ -6,14 +7,8 @@ const Messages = ({currentSlot, handleChange}) => {
         <Form.Group>
             <hr/>
             <Form.Label>Message <a className={'ms-2'} href={'https://docs.zmenu.dev/configurations/buttons#messages'} target={'_blank'}>(<i className="bi bi-question-lg"></i>)</a></Form.Label>
-            <Form.Control
-                name="messages"
-                as={'textarea'}
-                rows={10}
-                onChange={handleChange}
-                value={currentSlot.button?.messages ?? ''}
-                className={'rounded-1'}
-            />
+
+            <AutoCompleteFormControl key={'messages'} name={'messages'} asControl={'textarea'} handleChange={handleChange} defaultValue={currentSlot.button?.messages ?? ''} />
         </Form.Group>
     )