+++ changelog.md
@@ -1,9 +1,10 @@
 # Unreleased
 
-- You can now move around the inventory with the directional keys
-- Add inventories element in user show page
-- Add sound configuration
+- Add inventories element in user show page (admin)
+- Add messages, commands, console commands and sound configuration
 - Added a button to lock the page, to avoid scrolling and allow centered selections
+- Default slot selected will be slot 0
+- You can now move around the inventory with the directional keys
 
 # 1.0.3